+++ src/lib/components/modals/filters/AddNewButton.svelte
@@ -1,9 +1,10 @@
 <script lang="ts">
   import Link from '$components/ui/Link.svelte';
   import Plus from 'lucide-svelte/icons/plus';
-  import { filterStore } from '$stores/filters';
+  import { filterStore} from '$stores/filters';
   import { type Filter } from '$lib/filter';
   import { createEventDispatcher } from 'svelte';
+  import { nanoid } from 'nanoid';
 
   const dispatch = createEventDispatcher();
 
@@ -22,11 +23,11 @@
 
   function addNewFilter() {
     const newFilter = {
-      id: globalThis.crypto.randomUUID(),
+      id: nanoid(),
       name: getNextNewFilterName(),
       allowTags: [],
       blockTags: [],
-      enabled: false
+      onByDefault: false
     } as Filter;
 
     filterStore.update((filters) => [...filters, newFilter]);