+++ src/lib/components/ui/textarea/textarea.svelte
@@ -3,12 +3,15 @@
 	import type { TextareaEvents } from "./index.js";
 	import { cn } from "$lib/components/shadcn-utils.js";
 
-	type $$Props = HTMLTextareaAttributes;
+	type $$Props = HTMLTextareaAttributes & {
+		el?: HTMLTextAreaElement;
+	};
 	type $$Events = TextareaEvents;
 
 	let className: $$Props["class"] = undefined;
 	export let value: $$Props["value"] = undefined;
 	export { className as class };
+	export let el: $$Props["el"] = undefined;
 
 	// Workaround for https://github.com/sveltejs/svelte/issues/9305
 	// Fixed in Svelte 5, but not backported to 4.x.
@@ -17,7 +20,7 @@
 
 <textarea
 	class={cn(
-		"border-input bg-background ring-offset-background placeholder:text-muted-foreground focus-visible:ring-ring flex min-h-[80px] w-full rounded-lg border px-3 py-2 text-sm focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50",
+		"border-input bg-background ring-offset-background placeholder:text-muted-foreground focus-visible:ring-ring flex min-h-[80px] w-full rounded-lg border px-3 py-2 text-sm focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 invalid:border-error",
 		className
 	)}
 	bind:value
@@ -34,5 +37,6 @@
 	on:mouseleave
 	on:paste
 	on:input
+	bind:this={el}
 	{...$$restProps}
 ></textarea>